backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
SQLAlchemy>=2.0.21
python-dotenv>=1.0.0
pymongo>=4.6.0
google-genai>=1.0.0
python-multipart>=0.0.12
